+++ extensions/raindrop-io/src/components/BookmarkItem.tsx
@@ -61,7 +61,7 @@ export default function BookmarkItem(props: { bookmark: Bookmark; revalidate: ()
                 revalidate();
                 return res.json();
               } else {
-                throw new Error(res.statusText);
+                throw new Error("Error deleting link");
               }
             });
           } catch (error) {
@@ -80,54 +80,11 @@ export default function BookmarkItem(props: { bookmark: Bookmark; revalidate: ()
   const lastUpdatedDate = new Date(bookmark.lastUpdate);
   const createdDate = new Date(bookmark.created);
 
-  function subtitle() {
-    switch (preferences.additionalItemToDisplayInList) {
-      case "link":
-        return bookmark.link;
-      case "domain":
-        return bookmark.domain;
-      case "excerpt":
-        return bookmark.excerpt;
-      case "note":
-        return bookmark.note;
-    }
-  }
-
   function accessories() {
     const accessories = [];
 
     bookmark.tags.forEach((tag) => accessories.push({ tag: `#${tag}` }));
-
-    switch (bookmark.type) {
-      case "link":
-        accessories.push({ icon: Icon.Link });
-        break;
-      case "article":
-        accessories.push({ icon: Icon.FountainTip });
-        break;
-      case "image":
-        accessories.push({ icon: Icon.Image });
-        break;
-      case "video":
-        accessories.push({ icon: Icon.Video });
-        break;
-      case "audio":
-        accessories.push({ icon: Icon.Music });
-        break;
-      case "document":
-        accessories.push({ icon: Icon.Document });
-        break;
-    }
-
-    switch (preferences.displayDate) {
-      case "lastUpdated":
-        accessories.push({ date: lastUpdatedDate, tooltip: lastUpdatedDate.toLocaleString() });
-        break;
-      case "created":
-        accessories.push({ date: createdDate, tooltip: createdDate.toLocaleString() });
-        break;
-    }
-
+    accessories.push({ date: lastUpdatedDate, tooltip: lastUpdatedDate.toLocaleString() });
     return accessories;
   }
 
@@ -136,8 +93,7 @@ export default function BookmarkItem(props: { bookmark: Bookmark; revalidate: ()
       id={String(bookmark._id)}
       icon={getFavicon(bookmark.link, { fallback: "raindrop-icon.png" })}
       key={bookmark._id}
-      title={{ value: bookmark.title, tooltip: bookmark.link }}
-      subtitle={subtitle()}
+      title={bookmark.title}
       accessories={accessories()}
       actions={
         <ActionPanel>
@@ -163,8 +119,14 @@ export default function BookmarkItem(props: { bookmark: Bookmark; revalidate: ()
                 }
                 metadata={
                   <Detail.Metadata>
-                    <Detail.Metadata.Label title="Created" text={createdDate.toLocaleDateString()} />
-                    <Detail.Metadata.Label title="Last Updated" text={lastUpdatedDate.toLocaleDateString()} />
+                    <Detail.Metadata.Label
+                      title="Created"
+                      text={createdDate.toLocaleDateString()}
+                    />
+                    <Detail.Metadata.Label
+                      title="Last Updated"
+                      text={lastUpdatedDate.toLocaleDateString()}
+                    />
                     <Detail.Metadata.Label title="Domain" text={bookmark.domain} />
 
                     {bookmark.tags && (