+++ extensions/raindrop-io/src/hooks/useBrowserLink.ts
@@ -8,13 +8,21 @@ export function useBrowserLink() {
 
       switch (app.bundleId) {
         case "company.thebrowser.Browser":
-          return runAppleScript(`tell application "Arc" to return URL of active tab of front window`);
+          return runAppleScript(
+            `tell application "Arc" to return URL of active tab of front window`,
+          );
         case "com.vivaldi.Vivaldi":
-          return runAppleScript(`tell application "Vivaldi" to return URL of active tab of front window`);
+          return runAppleScript(
+            `tell application "Vivaldi" to return URL of active tab of front window`,
+          );
         case "com.google.Chrome":
-          return runAppleScript(`tell application "Google Chrome" to return URL of active tab of front window`);
+          return runAppleScript(
+            `tell application "Google Chrome" to return URL of active tab of front window`,
+          );
         case "com.brave.Browser":
-          return runAppleScript(`tell application "Brave Browser" to return URL of active tab of front window`);
+          return runAppleScript(
+            `tell application "Brave Browser" to return URL of active tab of front window`,
+          );
         case "com.apple.Safari":
           return runAppleScript(`tell application "Safari" to return URL of front document`);
         case "com.kagi.kagimacOS":
@@ -39,7 +47,9 @@ export function useBrowserLink() {
 
       // Fallback for Vivaldi Browser not recognized by bundleId
       if (app?.name === "Vivaldi.app") {
-        return runAppleScript(`tell application "Vivaldi" to return URL of active tab of front window`);
+        return runAppleScript(
+          `tell application "Vivaldi" to return URL of active tab of front window`,
+        );
       }
 
       throw new Error(`Unsupported App: ${app.name}`);