+++ extensions/raindrop-io/src/search.tsx
@@ -10,9 +10,13 @@ import { useLastUsedCollection } from "./hooks/useLastUsedCollection";
 
 export default function Main(): ReactElement {
   const preferences: Preferences = getPreferenceValues();
-  const [lastUsedCollection, setLastUsedCollection] = useCachedState<string>("last-used-collection", "0");
+  const [lastUsedCollection, setLastUsedCollection] = useCachedState<string>(
+    "last-used-collection",
+    "0",
+  );
 
-  const { getLastUsedCollection, setLastUsedCollection: setNextCollectionToUse } = useLastUsedCollection();
+  const { getLastUsedCollection, setLastUsedCollection: setNextCollectionToUse } =
+    useLastUsedCollection();
 
   useEffect(() => {
     const fetchLastUsedCollection = async () => {
@@ -25,7 +29,10 @@ export default function Main(): ReactElement {
   const defaultCollection = preferences.useLastCollection ? lastUsedCollection : "0";
 
   const [searchText, setSearchText] = useState<string>("");
-  const [collection, setCollection] = useCachedState<string>("selected-collection", defaultCollection);
+  const [collection, setCollection] = useCachedState<string>(
+    "selected-collection",
+    defaultCollection,
+  );
 
   const { isLoading, bookmarks, collections, revalidate } = useRequest({
     collection,
@@ -36,6 +43,15 @@ export default function Main(): ReactElement {
     if (collection !== value) {
       setCollection(value);
       setNextCollectionToUse(value);
+    }
+  };
+
+  const onSearch = (value: string) => {
+    if (value !== searchText) {
+      if (preferences.titleOnly) {
+        value = `title:"${value}"`;
+      }
+      setSearchText(value);
     }
   };
 
@@ -51,7 +67,7 @@ export default function Main(): ReactElement {
           defaultValue={collection}
         />
       }
-      onSearchTextChange={setSearchText}
+      onSearchTextChange={onSearch}
       isLoading={isLoading}
       throttle
     >