+++ extensions/raindrop-io/src/types.ts
@@ -44,12 +44,12 @@ export type Group = {
   sort: number;
   collections: number[];
 };
-
 export interface Preferences {
   token: string;
   useLastCollection?: boolean;
   additionalItemToDisplayInList?: string;
   displayDate?: string;
+  titleOnly?: boolean;
 }
 
 export interface CollectionsResponse {
@@ -100,7 +100,6 @@ export interface UserResponse {
 export interface CollectionItem {
   value?: number;
   label: string;
-  name?: string;
 }
 
 export type BookmarksParams = {
@@ -110,7 +109,6 @@ export type BookmarksParams = {
 
 export interface FormValues {
   link: string;
-  title?: string;
   collection?: string;
   tags: string[];
   newCollection?: string;